Fix Task 8 types
The Filter on "AdditionalLepton_idx" must be of type unsigned int.
However, it is compared with `-999`.  `-999` is used as a `null` value, to
show that the index was never updated.
Solution proposed is to initiliaze the `lep_idx` to 99999 which is certainly
larger than the largest possible `pt.size()`. Moreover, `best_i1` and
`best_i2` are also initialized to 99999, for the same reason so that
they are unsigned int.